# swagger.yaml
openapi: 3.0.3
info:
  title: Elasticsearch PoC API
  description: |
    A comprehensive Proof of Concept API for evaluating Elasticsearch as a search and analytics engine.
    
    This API provides endpoints for:
    - Product management (CRUD operations)
    - Advanced search capabilities with multiple query types
    - Analytics and aggregations
    - Performance testing and monitoring
    
    ## Features Tested
    - Full-text search with relevance scoring
    - Multi-field search across name, description, category
    - Fuzzy search with auto-correction
    - Wildcard and regex pattern matching
    - Boolean queries with must/should/filter clauses
    - Range filters for price, rating, dates
    - Faceted search with aggregation-based filters
    - Statistical and bucket aggregations
    - Performance benchmarking and load testing
    
  version: 1.0.0
  contact:
    name: Lawsikho Team
    email: support@lawsikho.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.example.com
    description: Production server (example)
...
paths:
...
  /api/products:
    get:
      tags:
        - Products
      summary: Get all products
      description: Retrieve products with pagination and sorting
      parameters:
        - name: page
          in: query
          description: Page number (1-based)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          description: Number of products per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: sort
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
            enum: [name, price, rating, createdAt, updatedAt]
            default: createdAt
        - name: order
          in: query
          description: Sort order
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  total:
                    type: integer
                    example: 1000
                  page:
                    type: integer
                    example: 1
                  size:
                    type: integer
                    example: 10
                  totalPages:
                    type: integer
                    example: 100
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Products
      summary: Create new product
      description: Create a new product in the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/search:
    post:
      tags:
        - Search
      summary: Advanced search
      description: Perform advanced search with multiple query types and filters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Search completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/search/suggest:
    get:
      tags:
        - Search
      summary: Get search suggestions
      description: Get autocomplete suggestions for search queries
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
            minLength: 1
        - name: field
          in: query
          description: Field to search in
          schema:
            type: string
            enum: [name, description, category]
            default: name
      responses:
        '200':
          description: Suggestions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  suggestions:
                    type: array
                    items:
                      type: object
                      properties:
                        text:
                          type: string
                          example: "smartphone"
                        score:
                          type: number
                          example: 0.95
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/search/facets:
    get:
      tags:
        - Search
      summary: Get search facets
      description: Get available facets for filtering search results
      parameters:
        - name: q
          in: query
          description: Optional search query to filter facets
          schema:
            type: string
      responses:
        '200':
          description: Facets retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacetsResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/search/related/{id}:
    get:
      tags:
        - Search
      summary: Find related products
      description: Find products related to a specific product
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
            format: uuid
        - name: size
          in: query
          description: Number of related products to return
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 5
      responses:
        '200':
          description: Related products found
          content:
            application/json:
              schema:
                type: object
                properties:
                  related:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  source:
                    $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Product:
      type: object
      required:
        - id
        - name
        - category
        - price
        - rating
        - inStock
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique product identifier
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          description: Product name
          example: "Smart Wireless Headphones"
        description:
          type: string
          description: Product description
          example: "High-quality wireless headphones with noise cancellation"
        category:
          type: string
          description: Product category
          example: "Electronics"
        price:
          type: number
          format: double
          description: Product price
          example: 299.99
        rating:
          type: number
          format: double
          minimum: 0
          maximum: 5
          description: Product rating
          example: 4.5
        tags:
          type: array
          items:
            type: string
          description: Product tags
          example: ["wireless", "premium", "noise-cancellation"]
        inStock:
          type: boolean
          description: Product availability
          example: true
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2024-01-01T00:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2024-01-01T00:00:00.000Z"
        metadata:
          type: object
          properties:
            brand:
              type: string
              example: "TechBrand"
            color:
              type: string
              example: "Black"
            size:
              type: string
              example: "M"
            weight:
              type: number
              format: double
              example: 1.5
        _score:
          type: number
          description: Search relevance score
          example: 0.95
    ProductInput:
      type: object
      required:
        - name
        - category
        - price
        - rating
        - inStock
      properties:
        name:
          type: string
          description: Product name
          example: "Smart Wireless Headphones"
        description:
          type: string
          description: Product description
          example: "High-quality wireless headphones with noise cancellation"
        category:
          type: string
          description: Product category
          example: "Electronics"
        price:
          type: number
          format: double
          description: Product price
          example: 299.99
        rating:
          type: number
          format: double
          minimum: 0
          maximum: 5
          description: Product rating
          example: 4.5
        tags:
          type: array
          items:
            type: string
          description: Product tags
          example: ["wireless", "premium", "noise-cancellation"]
        inStock:
          type: boolean
          description: Product availability
          example: true
        metadata:
          type: object
          properties:
            brand:
              type: string
              example: "TechBrand"
            color:
              type: string
              example: "Black"
            size:
              type: string
              example: "M"
            weight:
              type: number
              format: double
              example: 1.5
    SearchRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: Search query
          example: "smartphone"
        filters:
          type: object
          properties:
            category:
              type: string
              example: "Electronics"
            priceMin:
              type: number
              format: double
              example: 100
            priceMax:
              type: number
              format: double
              example: 500
            rating:
              type: number
              format: double
              example: 4.0
            inStock:
              type: boolean
              example: true
            tags:
              type: array
              items:
                type: string
              example: ["wireless", "premium"]
        page:
          type: integer
          minimum: 1
          default: 1
          description: Page number
        size:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
          description: Results per page
        sort:
          type: string
          enum: [name, price, rating, createdAt, updatedAt]
          default: createdAt
          description: Sort field
        order:
          type: string
          enum: [asc, desc]
          default: desc
          description: Sort order
        searchType:
          type: string
          enum: [multi_match, match_phrase, wildcard, fuzzy]
          default: multi_match
          description: Search query type
    SearchResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        total:
          type: integer
          description: Total number of results
          example: 150
        page:
          type: integer
          description: Current page number
          example: 1
        size:
          type: integer
          description: Results per page
          example: 10
        totalPages:
          type: integer
          description: Total number of pages
          example: 15
        searchType:
          type: string
          description: Search type used
          example: "multi_match"
        took:
          type: integer
          description: Search execution time in milliseconds
          example: 25
    FacetsResponse:
      type: object
      properties:
        facets:
          type: object
          properties:
            categories:
              type: array
              items:
                type: object
                properties:
                  key:
                    type: string
                    example: "Electronics"
                  doc_count:
                    type: integer
                    example: 150
            price_ranges:
              type: array
              items:
                type: object
                properties:
                  key:
                    type: string
                    example: "100-200"
                  doc_count:
                    type: integer
                    example: 75
            ratings:
              type: array
              items:
                type: object
                properties:
                  key:
                    type: string
                    example: "4-5"
                  doc_count:
                    type: integer
                    example: 100
            tags:
              type: array
              items:
                type: object
                properties:
                  key:
                    type: string
                    example: "wireless"
                  doc_count:
                    type: integer
                    example: 50
            in_stock:
              type: array
              items:
                type: object
                properties:
                  key:
                    type: string
                    example: "true"
                  doc_count:
                    type: integer
                    example: 120
        total:
          type: integer
          description: Total number of documents
          example: 1000
...
    Error:
      type: object
      required:
        - error
        - timestamp
      properties:
        error:
          type: string
          description: Error message
          example: "Internal Server Error"
        message:
          type: string
          description: Detailed error message
          example: "Something went wrong"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-01T00:00:00.000Z"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
security:
  - ApiKeyAuth: []